content/canvas/test/test_size.attributes.parse.badsuffix.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.parse.badsuffix</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100foo" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

todo(canvas.width == 100, "canvas.width == 100");

if (!_deferred) SimpleTest.finish();
});
</script>
